app_deepfake (2)/app_deepfake/requirements.txt
pillow==11.0.0
psycopg2-binary==2.9.10
